temp_package/requirements.txt
djangorestframework~=3.16.0
pip~=25.0.1
Django~=4.2.7
asgiref~=3.8.1
sqlparse~=0.5.3
numpy~=2.2.4
serializers~=0.2.4
typing_extensions~=4.13.0
setuptools~=65.5.0
future~=1.0.0
pytz~=2025.2
django-filter~=25.1